feat(test): reuse test-methods for Volvo and Saab
