+++ Saab95Test.java
@@ -1,0 +1,7 @@
+class Saab95Test extends GenericCarTest {
+
+    @Override
+    protected Car createCar() {
+        return new Saab95();
+    }
+}
